refactor api test
